requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-storages==1.14.2
django-widget-tweaks==1.5.0
dropbox==11.36.2
idna==3.6
pillow==10.2.0
ply==3.11
psycopg==3.1.17
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Markdown==3.7
